Feat: 등록하기 컴포넌트 분리
